WIP to modularise routers
